Biggest update in mewdekos history. Moving from efcore to linqtodb.
(cherry picked from commit 0d7ae94168ef786091f296370460a5d694008284)

+++ src/Mewdeko/Modules/Utility/AiCommands.cs
@@ -62,7 +62,7 @@ public partial class Utility
                 return;
             }
 
-            config.Provider = provider;
+            config.Provider = (int)provider;
             config.Model = model;
             await Service.UpdateConfig(config);
